Add authorization guards. Implement discover services.

+++ SOA/soa-backend/src/answers/answers.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectEntityManager } from '@nestjs/typeorm';
 import { Question } from 'src/questions/entities/question.entity';
 import { QuestionsService } from 'src/questions/questions.service';
@@ -17,12 +17,13 @@ export class AnswersService {
 
     // Create new answer
     // Check wether user and question exist. If so, create new answer.
-    async createAnswer(createAnswerDto: CreateAnswerDto): Promise<Answer> {
+    async createAnswer(createAnswerDto: CreateAnswerDto, user): Promise<Answer> {
         return this.manager.transaction(async manager => {
             // findUserFromEmail already takes care of the exception if needed!
-            await this.usersService.findUserFromEmail(createAnswerDto.createdBy);
+            // await this.usersService.findUserFromEmail(createAnswerDto.createdBy);
             const questionExists = await this.questionsService.findOneQuestion(createAnswerDto.questionId);
-            if(!questionExists) throw new NotFoundException('Question does not exits. Answer not created!')
+            if(!questionExists) throw new NotFoundException('Question does not exits. Answer not created!');
+            if(user.email != createAnswerDto.createdBy) throw new ConflictException('User cannot create answer for another email');
             const newAnswer = await manager.create(Answer, createAnswerDto);
             newAnswer.question = questionExists;
             return await manager.save(newAnswer);
@@ -42,25 +43,27 @@ export class AnswersService {
     }
 
     // Get Users Answers
-    async findUserAnswers(updateAnswerDto: UpdateAnswerDto): Promise<Answer[]> {
+    async findUserAnswers(updateAnswerDto: UpdateAnswerDto, user): Promise<Answer[]> {
         return this.manager.transaction(async manager => {
             // findUserFromEmail already takes care of the exception if needed!
-            if(!updateAnswerDto.createdBy) throw new BadRequestException('Please provide user email')
-            await this.usersService.findUserFromEmail(updateAnswerDto.createdBy);
+            if(!updateAnswerDto.createdBy) throw new BadRequestException('Please provide user email');
+            if(user.email != updateAnswerDto.createdBy) throw new ConflictException('User cannot search questions created by another account from here')
+            // await this.usersService.findUserFromEmail(updateAnswerDto.createdBy);
             const userQuestions = manager.find(Answer, {where: {createdBy: updateAnswerDto.createdBy}, relations: ['question', 'question.labels']}) 
             return userQuestions;
         })
     }
 
     // Update answer
-    async updateAnswer(answerId: number, updateAnswerDto: UpdateAnswerDto): Promise<Answer> {
+    async updateAnswer(answerId: number, updateAnswerDto: UpdateAnswerDto, user): Promise<Answer> {
         return this.manager.transaction(async manager => {
             const answerExists = await manager.findOne(Answer, answerId, {relations: ['question']});
+            if(updateAnswerDto.createdBy != undefined && user.email != updateAnswerDto ) throw new ConflictException('User can only create or update his own answers')
             if(!answerExists) throw new NotFoundException('Answer does not exist!');
-            if(updateAnswerDto.createdBy != answerExists.createdBy){
-                // findUserFromEmail already takes care of the exception if needed!
-                await this.usersService.findUserFromEmail(updateAnswerDto.createdBy);
-            }
+            // if(updateAnswerDto.createdBy != answerExists.createdBy){
+            //     // findUserFromEmail already takes care of the exception if needed!
+            //     await this.usersService.findUserFromEmail(updateAnswerDto.createdBy);
+            // }
             if(updateAnswerDto.questionId != answerExists.question.questionId){
                 const newQuestionExists = await manager.findOne(Question, updateAnswerDto.questionId);
                 if(!newQuestionExists) throw new NotFoundException('New question does not exits. Answer not updated!');
@@ -72,10 +75,11 @@ export class AnswersService {
     }
 
     // Delete answer
-    async removeAnswer(answerId: number):Promise<any> {
+    async removeAnswer(answerId: number, user):Promise<any> {
         return this.manager.transaction(async manager => {
             const answerExists = await this.manager.findOne(Answer, answerId, {relations: ['question']});
             if(!answerExists) throw new NotFoundException('Question does not exits!');
+            if(user.email != answerExists.createdBy) throw new ConflictException('User cannot delete questions created by another user')
             return await manager.delete(Answer, answerId);  
         })
     }
